requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
crypto==1.4.1
decorator==4.4.2
idna==2.9
ijson==2.6.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
ldap3==2.7
mysql-connector==2.2.9
Naked==0.1.31
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycrypto==2.6.1
Pygments==2.6.1
pysnow==0.7.14
python-magic==0.4.15
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
shellescape==3.8.1
six==1.14.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
